Fix issues found with match nara unit tests
new_risk_spreadsheet() and match_nara_risk() need nara_df as a parameter for unit testing to work.

In match_nara_risk(), removing version of "NO VALUE" for name_version column was not working, added NARA_File_Extensions to the result, and fixed Technique 2 to use lowercase versions of columns.